Remove Node runtime fallback
