+++ .env.template
@@ -3,7 +3,3 @@ TOKEN=
 
 # Optional token for Tito-related utility scripts.
 TITO_TOKEN=<get from https://api.tito.io/>
-
-# Optional comma-separated list of webhook URLs to POST after the Node server
-# has started in production.
-REBUILD_SITES=
